pyproject.toml
[project]
name = "speed-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "genanki>=0.13.1",
    "google-genai>=1.57.0",
    "python-dotenv>=1.2.1",
]
